Moved CompressedObjectParcelable into adapters.data
