+++ app/src/main/java/com/amaze/filemanager/asynchronous/asynctasks/RarHelperTask.java
@@ -3,7 +3,7 @@ package com.amaze.filemanager.asynchronous.asynctasks;
 import android.os.AsyncTask;
 
 import com.amaze.filemanager.filesystem.compressed.RarHelper;
-import com.amaze.filemanager.ui.CompressedObjectParcelable;
+import com.amaze.filemanager.adapters.data.CompressedObjectParcelable;
 import com.amaze.filemanager.utils.OnAsyncTaskFinished;
 import com.github.junrar.Archive;
 import com.github.junrar.exception.RarException;
